requirements.txt
outline-vpn-api==6.2.0
python-telegram-bot==21.3
click==8.1.7
python-dotenv==1.0.0
